feat: save aggregated contents and read it in api
